+++ Changelog.md
@@ -1,6 +1,7 @@
 # 11.2.5-20251008-1
 * Updated data files with latest data from 11.2.5 week 2025-41
 * Cleanup code in ObjectiveTracker.
+* Sort the pet list in the objective tracker.
 
 # 11.2.5-20251006-1
 * Updated TOC to match WoW 11.2.5